Move some utility methods to a separated util file in Home component
