+++ src/components/screens/tabs/home/index.js
@@ -11,25 +11,21 @@ import {
   accountFetched as accountFetchedAction,
 } from '../../../../actions/accounts';
 import { settingsUpdated as settingsUpdatedAction } from '../../../../actions/settings';
-import ModalHolder from '../../../../utilities/modal';
 import AccountSummary from './accountSummary/home';
 import Transactions from '../../../shared/transactions';
 import Empty from '../../../shared/transactions/empty';
 import Loading from '../../../shared/transactions/loading';
-import IntroModal from './introModal';
+import { loadMore, resetTxAndFetch, showIntroModal, showInitializationModal } from './utils';
 import { viewportHeight } from '../../../../utilities/device';
 import InfiniteScrollView from '../../../shared/infiniteScrollView';
 import { tokenMap, tokenKeys } from '../../../../constants/tokens';
 import withTheme from '../../../shared/withTheme';
 import getStyles from './styles';
 import { themes } from '../../../../constants/styleGuide';
-import { fromRawLsk } from '../../../../utilities/conversions';
-import InitializationModal from './initializationModal';
 import HomeHeaderTitle from '../../router/homeHeaderTitle';
 
 const itemHeight = 90;
 const summaryHeight = 200;
-
 /**
  * This component would be mounted first and would be used to config and redirect
  * the application to referer page or Sign In
@@ -85,7 +81,6 @@ class Home extends React.Component {
       account,
       incognito,
     } = this.props;
-
     setParams({
       title: {
         placeHolder: `${tokenMap[activeToken].label} Wallet`,
@@ -112,25 +107,6 @@ class Home extends React.Component {
     });
   };
 
-  resetTxAndFetch = () => {
-    const {
-      transactionsReset,
-      transactionsLoaded,
-      account,
-      activeToken,
-    } = this.props;
-
-    transactionsReset();
-
-    // giving some time for the transition animations to settle
-    this.initialFetchTimeout = setTimeout(() => {
-      transactionsLoaded({
-        address: account[activeToken].address,
-        offset: 0,
-      });
-    }, 200);
-  };
-
   onScroll() {
     return Animated.event([
       {
@@ -139,66 +115,18 @@ class Home extends React.Component {
     ]);
   }
 
-  loadMore = () => {
-    const {
-      activeToken,
-      account,
-      transactionsLoaded,
-      transactions,
-    } = this.props;
-
-    if (account[activeToken]) {
-      transactionsLoaded({
-        address: account[activeToken].address,
-        offset: transactions.confirmed.length,
-      });
-    }
-  };
-
   refreshAccountAndTx = () => {
     this.lastActiveToken = this.props.activeToken;
-    this.resetTxAndFetch();
+    this.initialFetchTimeout = setTimeout(() => { resetTxAndFetch(this.props) }, 200);
     this.props.accountFetched();
-  };
-
-  showIntroModal = () => {
-    if (!this.props.btcIntroShown) {
-      this.modalTimeout = setTimeout(() => {
-        ModalHolder.open({
-          title: 'We’ve got good news!',
-          component: IntroModal,
-        });
-        this.props.settingsUpdated({ btcIntroShown: true });
-      }, 1200);
-    }
-  };
-
-  showInitializationModal = () => {
-    const { account, activeToken, transactions } = this.props;
-    const balance = parseFloat(fromRawLsk(account[tokenMap.LSK.key].balance));
-
-    if (
-      activeToken === tokenMap.LSK.key &&
-      !account[activeToken].initialized &&
-      (!transactions || transactions.pending.length < 1) &&
-      balance >= 0.2
-    ) {
-      ModalHolder.open({
-        title: 'Initialize your account',
-        component: InitializationModal,
-        callback: () =>
-          this.props.navigation.navigate('Send', { initialize: true }),
-      });
-    }
   };
 
   screenWillFocus = () => {
     if (this.lastActiveToken === null) {
       this.bindInfiniteScroll();
       this.setHeader();
-      this.showIntroModal();
+      this.modalTimeout = setTimeout(() => { showIntroModal(this.props) }, 1200);
     }
-
     if (this.lastActiveToken !== this.props.activeToken) {
       this.refreshAccountAndTx();
       this.setHeader();
@@ -213,19 +141,15 @@ class Home extends React.Component {
       incognito,
     } = this.props;
     addListener('willFocus', this.screenWillFocus);
-
     if (state.params && state.params.discreet && !incognito) {
       settingsUpdated({ incognito: true });
     }
-
     navigation.setParams({ scrollToTop: this.scrollToTop });
-
     this.accountFetchTimeout = setTimeout(() => {
       this.fetchInactiveTokensAccounts();
     }, 1000);
-
     setTimeout(() => {
-      this.showInitializationModal();
+      showInitializationModal(this.props);
     }, 1200);
   }
 
@@ -241,14 +165,11 @@ class Home extends React.Component {
       },
     } = this.props;
     const prevTokenList = prevProps.settings.token.list;
-
     const prevTransactionCount =
       prevProps.transactions.pending.length +
       prevProps.transactions.confirmed.length;
-
     const transactionCount =
       transactions.pending.length + transactions.confirmed.length;
-
     const shouldUpdateState =
       prevProps.transactions.loaded !== transactions.loaded ||
       prevTransactionCount !== transactionCount;
@@ -260,18 +181,15 @@ class Home extends React.Component {
           transactionCount,
       });
     }
-
     if (this.shouldFetchAccounts(prevTokenList, list)) {
       this.fetchInactiveTokensAccounts();
     }
-
     if (
       prevProps.account[activeToken].balance !== account[activeToken].balance ||
       prevProps.incognito !== incognito
     ) {
       this.setHeader();
     }
-
     if (prevProps.activeToken !== activeToken && isFocused) {
       this.refreshAccountAndTx();
       this.setHeader();
@@ -292,7 +210,6 @@ class Home extends React.Component {
     const inactiveTokens = tokenKeys.filter(
       key => settings.token.list[key] && key !== activeToken
     );
-
     if (inactiveTokens.length > 0) {
       inactiveTokens.forEach(token => {
         accountFetched(token);
@@ -311,9 +228,7 @@ class Home extends React.Component {
       isFocused,
       activeToken,
     } = this.props;
-
     let content = null;
-
     if (!transactions.loaded) {
       content = <Loading style={styles.loadingState} />;
     } else {
@@ -321,7 +236,6 @@ class Home extends React.Component {
         transactions.count > 0
           ? [...transactions.pending, ...transactions.confirmed]
           : ['emptyState'];
-
       content = (
         <InfiniteScrollView
           ref={el => {
@@ -331,7 +245,7 @@ class Home extends React.Component {
           onScroll={this.onScroll.call(this)}
           style={[styles.scrollView]}
           refresh={updateTransactions}
-          loadMore={this.loadMore}
+          loadMore={() => { loadMore(this.props); }}
           list={listElements}
           count={transactions.count}
           render={refreshing =>
